Updating figures

+++ common/quaternion.py
@@ -7,7 +7,6 @@
 
 import torch
 import numpy as np
-
 # PyTorch-backed implementations
 
 def qmul(q, r):
@@ -18,9 +17,9 @@ def qmul(q, r):
     """
     assert q.shape[-1] == 4
     assert r.shape[-1] == 4
-    
+
     original_shape = q.shape
-    
+
     # Compute outer product
     terms = torch.bmm(r.view(-1, 4, 1), q.view(-1, 1, 4))
 
@@ -40,11 +39,11 @@ def qrot(q, v):
     assert q.shape[-1] == 4
     assert v.shape[-1] == 3
     assert q.shape[:-1] == v.shape[:-1]
-    
+
     original_shape = list(v.shape)
     q = q.view(-1, 4)
     v = v.view(-1, 3)
-    
+
     qvec = q[:, 1:]
     uv = torch.cross(qvec, v, dim=1)
     uuv = torch.cross(qvec, uv, dim=1)
@@ -57,16 +56,16 @@ def qeuler(q, order, epsilon=0):
     Returns a tensor of shape (*, 3).
     """
     assert q.shape[-1] == 4
-    
+
     original_shape = list(q.shape)
     original_shape[-1] = 3
     q = q.view(-1, 4)
-    
+
     q0 = q[:, 0]
     q1 = q[:, 1]
     q2 = q[:, 2]
     q3 = q[:, 3]
-    
+
     if order == 'xyz':
         x = torch.atan2(2 * (q0 * q1 - q2 * q3), 1 - 2*(q1 * q1 + q2 * q2))
         y = torch.asin(torch.clamp(2 * (q1 * q3 + q0 * q2), -1+epsilon, 1-epsilon))
@@ -104,8 +103,8 @@ def qmul_np(q, r):
     return qmul(q, r).numpy()
 
 def qrot_np(q, v):
-    q = torch.from_numpy(q).contiguous()
-    v = torch.from_numpy(v).contiguous()
+    q = torch.from_numpy(q).contiguous().double()
+    v = torch.from_numpy(v).contiguous().double()
     return qrot(q, v).numpy()
 
 def qeuler_np(q, order, epsilon=0, use_gpu=False):
@@ -121,13 +120,13 @@ def qfix(q):
     Enforce quaternion continuity across the time dimension by selecting
     the representation (q or -q) with minimal distance (or, equivalently, maximal dot product)
     between two consecutive frames.
-    
+
     Expects a tensor of shape (L, J, 4), where L is the sequence length and J is the number of joints.
     Returns a tensor of the same shape.
     """
     assert len(q.shape) == 3
     assert q.shape[-1] == 4
-    
+
     result = q.copy()
     dot_products = np.sum(q[1:]*q[:-1], axis=2)
     mask = dot_products < 0
@@ -143,7 +142,7 @@ def expmap_to_quaternion(e):
     Returns a tensor of shape (*, 4).
     """
     assert e.shape[-1] == 3
-    
+
     original_shape = list(e.shape)
     original_shape[-1] = 4
     e = e.reshape(-1, 3)
@@ -158,16 +157,16 @@ def euler_to_quaternion(e, order):
     Convert Euler angles to quaternions.
     """
     assert e.shape[-1] == 3
-    
+
     original_shape = list(e.shape)
     original_shape[-1] = 4
-    
+
     e = e.reshape(-1, 3)
-    
+
     x = e[:, 0]
     y = e[:, 1]
     z = e[:, 2]
-    
+
     rx = np.stack((np.cos(x/2), np.sin(x/2), np.zeros_like(x), np.zeros_like(x)), axis=1)
     ry = np.stack((np.cos(y/2), np.zeros_like(y), np.sin(y/2), np.zeros_like(y)), axis=1)
     rz = np.stack((np.cos(z/2), np.zeros_like(z), np.zeros_like(z), np.sin(z/2)), axis=1)
@@ -186,10 +185,9 @@ def euler_to_quaternion(e, order):
             result = r
         else:
             result = qmul_np(result, r)
-            
+
     # Reverse antipodal representation to have a non-negative "w"
     if order in ['xyz', 'yzx', 'zxy']:
         result *= -1
-    
+
     return result.reshape(original_shape)
-
